Use Enum for worker status.
Looking at the code it was hard to find all the states a worker could be
in, and in particular it is easy to forget 'closing_gracefully' for
example when skimming over a `self.status.startswith('clos')`.

So I just took a pass at seeing in how many places status was
assigned/compared in worker and scheduler and replace that by an Enum
(I'm assuming the status each can see is shared though I did not found
mention of graceful shutdown in scheduler, and it looks like status
_can_ be `None` in one place.

Note that even is Enums have values for their variants, the variants _do
not_ compare equal to their values. That is to say `Status.running =!
'running'`, but `Status.running.value == 'running'`, so this woudl
technically be an API changes for anything that directly manipulating
worker status, and  subclasses.