Use Enum for worker status.
Looking at the code it was hard to find all the states a worker could be
in, and in particular it is easy to forget 'closing_gracefully' for
example when skimming over a `self.status.startswith('clos')`, and it
also hard to fine that in some places the status can be "None", "init",
"starting" and "stopped"

So I just took a pass at seeing in how many places status was
assigned/compared in worker, nanny and scheduler and replace that by an Enum
I'm assuming the status each can see is shared though I did not found
mention of graceful shutdown in scheduler, nor of "init" outside Nanny
for example.

I tried to do separate Status Enums for Scheduler/Nanny/Worker though it
appeared to fail in some area.

Note that even is Enums have values for their variants, the variants _do
not_ compare equal to their values. That is to say `Status.running =!
'running'`, but `Status.running.value == 'running'`, so this would
technically be an API changes for anything that directly manipulating
worker status, and  subclasses.

I've overridden __eq__ to raise in that case, but it's mostly to check
if I've forgotten places, it woudl of course we updates to allow
Enum/str comparison