Avoid too aggressive retry of connections.
Originally retrying was implemented to avoid timeout expiring for
addresses not yet existing. Though giving up after 1 second maximum can
be annoying for slow system and can trigger a thundering herd-like
problem.

Here progressively increase the inner timeout to give a chance to a slow
systems on th other side to respond especially when they are slow.

Here also jitter a bit when we are going to retry, and for how long
before we give up in the inner timeout to try-to again mitigate teh
thundering herd.

This is relatively difficult to test; so no tests are included.